requirements.txt
pandas
numpy
matplotlib
python-pptx